+++ Nesting_4/Base.cs
@@ -9,42 +9,42 @@ namespace Nesting_4
         /// <summary>
         /// coordinata P in basso a sinitra della green area
         /// </summary>
-        public float BLpPosition { get; set; } = 0;
+        public double BLpPosition { get; set; } = 0;
 
         /// <summary>
         /// coordinata Q in basso a sinistra della green area
         /// </summary>
-        public float BLqPosition { get; set; } = 0;
+        public double BLqPosition { get; set; } = 0;
 
         /// <summary>
         /// coordinata P in alto a sinistra della green area
         /// </summary>
-        public float TLpPosition { get; set; } = 0;
+        public double TLpPosition { get; set; } = 0;
 
         /// <summary>
         /// coordinata Q in alto a sinistra della green area
         /// </summary>
-        public float TLqPosition { get; set; } = 0;
+        public double TLqPosition { get; set; } = 0;
 
         /// <summary>
         /// coordinata in basso a destra della green area
         /// </summary>
-        public float BRpPosition { get; set; } = 0;
+        public double BRpPosition { get; set; } = 0;
 
         /// <summary>
         /// coordinata in basso a destra della green area
         /// </summary>
-        public float BRqPosition { get; set; } = 0;
+        public double BRqPosition { get; set; } = 0;
 
         /// <summary>
         /// coordinata in basso a destra della green area
         /// </summary>
-        public float TRpPosition { get; set; } = 0;
+        public double TRpPosition { get; set; } = 0;
 
         /// <summary>
         /// coordinata in basso a destra della green area
         /// </summary>
-        public float TRqPosition { get; set; } = 0;
+        public double TRqPosition { get; set; } = 0;
 
     }
 }
